rag_backend/requirements.txt
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
fastapi>=0.104.0,<0.105.0
pydantic>=2.0.0,<3.0.0
uvicorn[standard]
PyPDF2>=2.0.0
